--- bindings/nodejs/lib/types/models/api/node/routes-response.ts
@@ -1,12 +1,0 @@
-// Copyright 2023 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-/**
- * Response from the /routes endpoint.
- */
-export interface RoutesResponse {
-    /**
-     * The routes the node exposes.
-     */
-    routes: string[];
-}
