+++ bindings/nodejs/lib/types/models/api/response.ts
@@ -1,21 +1,22 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-/**
- * Base response data.
- */
-export interface Response {
-    /**
-     * Optional error in the response.
-     */
-    error?: {
-        /**
-         * The code for the error response.
-         */
-        code: string;
-        /**
-         * A more descriptive version of the error.
-         */
-        message: string;
-    };
-}
+// TODO: can this be removed? Nothing seems to require it :thinking:
+// /**
+//  * Base response data.
+//  */
+// export interface Response {
+//     /**
+//      * Optional error in the response.
+//      */
+//     error?: {
+//         /**
+//          * The code for the error response.
+//          */
+//         code: string;
+//         /**
+//          * A more descriptive version of the error.
+//          */
+//         message: string;
+//     };
+// }
