+++ bindings/nodejs/lib/types/models/api/utxo.ts
@@ -1,5 +1,265 @@
-// Copyright 2024 IOTA Stiftung
+// Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
+
+// Temp solution for not double parsing JSON
+import { plainToInstance, Type } from 'class-transformer';
+import { BlockId, TransactionId } from '../../block/id';
+import { Output, OutputId, OutputDiscriminator } from '../../block/output';
+import { SlotCommitmentId, SlotIndex } from '../../block/slot';
+import { TransactionState } from '../state';
+import { HexEncodedString } from '../../utils';
+
+/**
+ * OutputCommitmentProof types.
+ */
+export enum OutputCommitmentProofType {
+    /** Denotes a HashableNode. */
+    HashableNode = 0,
+    /** Denotes a LeafHash. */
+    LeafHash = 1,
+    /** Denotes a ValueHash. */
+    ValueHash = 2,
+}
+
+export abstract class OutputCommitmentProof {
+    readonly type: OutputCommitmentProofType;
+
+    /**
+     * @param type The type of OutputCommitmentProof.
+     */
+    constructor(type: OutputCommitmentProofType) {
+        this.type = type;
+    }
+
+    /**
+     * Parse an OutputCommitmentProof from a plain JSON object.
+     */
+    public static parse(data: any): OutputCommitmentProof {
+        if (data.type == OutputCommitmentProofType.HashableNode) {
+            return plainToInstance(HashableNode, data) as any as HashableNode;
+        } else if (data.type == OutputCommitmentProofType.LeafHash) {
+            return plainToInstance(LeafHash, data) as any as LeafHash;
+        } else if (data.type == OutputCommitmentProofType.ValueHash) {
+            return plainToInstance(ValueHash, data) as any as ValueHash;
+        }
+        throw new Error('Invalid JSON');
+    }
+}
+
+/**
+ * Contains the hashes of the left and right children of a node in the OutputCommitmentProof tree.
+ */
+export class HashableNode extends OutputCommitmentProof {
+    readonly l: OutputCommitmentProof;
+    readonly r: OutputCommitmentProof;
+
+    /**
+     * @param l Output commitment proof of the left subtree.
+     * @param r Output commitment proof of the right subtree.
+     */
+    constructor(l: OutputCommitmentProof, r: OutputCommitmentProof) {
+        super(OutputCommitmentProofType.HashableNode);
+        this.l = l;
+        this.r = r;
+    }
+}
+
+/**
+ * Contains the hash of a leaf in the OutputCommitmentProof tree.
+ */
+export class LeafHash extends OutputCommitmentProof {
+    readonly hash: HexEncodedString;
+
+    /**
+     * @param hash The hash of the leaf.
+     */
+    constructor(hash: HexEncodedString) {
+        super(OutputCommitmentProofType.LeafHash);
+        this.hash = hash;
+    }
+}
+
+/**
+ * Contains the hash of the value for which the OutputCommitmentProof is being computed.
+ */
+export class ValueHash extends OutputCommitmentProof {
+    readonly hash: HexEncodedString;
+
+    /**
+     * @param hash The hash of the value.
+     */
+    constructor(hash: HexEncodedString) {
+        super(OutputCommitmentProofType.ValueHash);
+        this.hash = hash;
+    }
+}
+
+/**
+ * The proof of the output identifier.
+ */
+export class OutputIdProof {
+    readonly slot: SlotIndex;
+    readonly outputIndex: number;
+    readonly transactionCommitment: HexEncodedString;
+    readonly outputCommitmentProof: OutputCommitmentProof;
+
+    /**
+     * @param slot The slot index of the output.
+     * @param outputIndex The index of the output within the corresponding transaction.
+     * @param transactionCommitment The commitment of the transaction that created the output. Hex-encoded with 0x prefix.
+     * @param outputCommitmentProof The proof of the output commitment.
+     */
+    constructor(
+        slot: SlotIndex,
+        outputIndex: number,
+        transactionCommitment: HexEncodedString,
+        outputCommitmentProof: OutputCommitmentProof,
+    ) {
+        this.slot = slot;
+        this.outputIndex = outputIndex;
+        this.transactionCommitment = transactionCommitment;
+        this.outputCommitmentProof = outputCommitmentProof;
+    }
+}
+
+/**
+ * An output with its output id proof.
+ * Response of GET /api/core/v3/outputs/{outputId}.
+ */
+export class OutputResponse {
+    /**
+     * One of the possible output types.
+     */
+    @Type(() => Output, {
+        discriminator: OutputDiscriminator,
+    })
+    output!: Output;
+    /**
+     * The proof of the output identifier.
+     */
+    outputIdProof!: OutputIdProof;
+}
+
+/**
+ * Metadata of the output if it is included in the ledger.
+ */
+export interface OutputInclusionMetadata {
+    /**
+     * Slot in which the output was included.
+     */
+    slot: SlotIndex;
+    /**
+     * Transaction ID that created the output.
+     */
+    transactionId: TransactionId;
+    /**
+     * Commitment ID that includes the creation of the output.
+     */
+    commitmentId?: SlotCommitmentId;
+}
+
+/**
+ * Metadata of the output if it is marked as spent in the ledger.
+ */
+export interface OutputConsumptionMetadata {
+    /**
+     * Slot in which the output was spent.
+     */
+    slot: SlotIndex;
+    /**
+     * Transaction ID that spent the output.
+     */
+    transactionId: TransactionId;
+    /**
+     * Commitment ID that includes the spending of the output.
+     */
+    commitmentId?: SlotCommitmentId;
+}
+
+/**
+ * Metadata of an output.
+ */
+export interface OutputMetadataResponse {
+    /**
+     * The ID of the output.
+     */
+    outputId: OutputId;
+    /**
+     * The ID of the block in which the output was included.
+     */
+    blockId: BlockId;
+    /**
+     * Metadata of the output if it is included in the ledger.
+     */
+    included: OutputInclusionMetadata;
+    /**
+     * Metadata of the output if it is marked as spent in the ledger.
+     */
+    spent?: OutputConsumptionMetadata;
+    /**
+     * Latest commitment ID of the node.
+     */
+    latestCommitmentId: SlotCommitmentId;
+}
+
+/**
+ * An output with its output id proof and its metadata.
+ * Response of GET /api/core/v3/outputs/{outputId}/full.
+ */
+export class OutputWithMetadataResponse {
+    /**
+     * One of the possible output types.
+     */
+    output!: Output;
+    /**
+     * The associated Output ID proof.
+     */
+    outputIdProof!: OutputIdProof;
+    /**
+     * The metadata of the output.
+     */
+    metadata!: OutputMetadataResponse;
+}
+
+/**
+ * An output and its metadata.
+ */
+export class OutputWithMetadata {
+    /**
+     * One of the possible output types.
+     */
+    output!: Output;
+    /**
+     * The metadata of the output.
+     */
+    metadata!: OutputMetadataResponse;
+}
+
+/**
+ * Metadata of a transaction.
+ */
+export interface TransactionMetadataResponse {
+    /**
+     * The transaction id.
+     */
+    transactionId: TransactionId;
+    /**
+     * The transaction state.
+     */
+    transactionState: TransactionState;
+    /**
+     * The slot of the earliest included valid block that contains an attachment of the transaction.
+     */
+    earliestAttachmentSlot: SlotIndex;
+    /**
+     * If applicable, indicates the error that occurred during the transaction processing.
+     */
+    transactionFailureReason?: TransactionFailureReason;
+    /**
+     * Contains the detailed error message that occurred during the transaction processing if the debug mode was activated in the retainer.
+     */
+    transactionFailureDetails?: string;
+}
 
 /**
  * Reason for transaction failure.
