--- bindings/nodejs/lib/types/models/api/utxo/index.ts
@@ -1,8 +1,0 @@
-// Copyright 2024 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-export * from './output-id-proof';
-export * from './output-metadata-response';
-export * from './output-response';
-export * from './transaction-failure-reason';
-export * from './transaction-response';
