--- bindings/nodejs/lib/types/models/api/validators/index.ts
@@ -1,4 +1,0 @@
-// Copyright 2024 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-export * from './validators-response';
